test: preserve existing branding colors to prevent Authentication Profile reset
